+++ src/constants.js
@@ -1,5 +1,5 @@
-export const AI_WRAPPED_KEY = Symbol("AI_WRAPPED_KEY");
+export const AI_PATCHED_KEY = Symbol("AI_PATCHED_KEY");
 
-export const MIDDLEWARE = "Middleware";
-export const PARAM_MIDDLEWARE = "Param Middleware";
-export const ERROR_MIDDLEWARE = "Error Middleware";
+export const MIDDLEWARE = "Express Middleware";
+export const PARAM_MIDDLEWARE = "Express Param Middleware";
+export const ERROR_MIDDLEWARE = "Express Error Middleware";