+++ src/isPatched.js
@@ -1,0 +1,5 @@
+import { AI_PATCHED_KEY } from "./constants";
+
+const isWrapped = (objectLike) => objectLike[AI_PATCHED_KEY];
+
+export default isWrapped;