+++ src/isWrapped.js
@@ -1,5 +1,0 @@
-import { AI_WRAPPED_KEY } from "./constants";
-
-const isWrapped = (objectLike) => objectLike[AI_WRAPPED_KEY] === true;
-
-export default isWrapped;