+++ src/usePatch/joinPaths.js
@@ -1,0 +1,5 @@
+import path from "path";
+
+const joinPaths = (basePath, pathSegment) => path.join(basePath, pathSegment);
+
+export default joinPaths;